Backend/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.1
protobuf==4.21.6
PyYAML==6.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2